requirements.txt
psycopg2-binary
redis
pymongo
python-dotenv
pandas
scikit-learn
joblib
numpy
fastapi
uvicorn
xgboost
